fix: Call app.setLoading and load pattern constructor. Fixes #547
The fact that we were calling setLoading() instead of app.setLoading()
is a straight-forward fix, but the issue with the constructor not so much.

We sometimes pass the pattern constructor to the export component (when
we load it from the draft UI) and sometimes we do not (when we load it
from the pattern page). That's not great, but we're seeing a lot of
users run imto this bug so I've fixed it by loading the constructor
in the export component if it's not passed.

This should get refactored as we are re-working the entire save/export
situation anyway.