chore: Removed contributors placeholder content
I initially populated the contributors file with a number of people
to showcase the new design.
Now, I've removed them all and will be adding them again but only
if they sign off on it. It's a way of opting in with a papertrail.
So I will create a merge request for each contributor that needs
to be approved by the person.