chore: Added @annekecaramin as a contributor
This commit will add @annekecaramin to the contributors file.
The result is that she will be listed on the website as a contributor.
For this reason, @annekecaramin should approve this before this gets
merged.

In other words, this commit is the papertrail for the opt-in way of
adding contributors to the site.