chore: Added @dfbean as a contributor
This commit will add @dfbean to the contributors file.
The result is that she will be listed on the website as a contributor.
For this reason, @dfbean should approve this before this gets
merged.

In other words, this commit is the papertrail for the opt-in way of
adding contributors to the site.

Notes:

 - I used `Deb` here as a name
 - I did not add a picture only because that seems to be what you
 prefer