refactor: Build script, context, and app hook
This wraps up a rather large refactor of the frontend code base.
It boils down to three things:

**Move MDX processing to the build script**

This applies to docs, blogs and showcases, but is most relevant for docs.
We use to have a component that would have a static graphql query to
pull up all docs info. On every load of a documentation page, it would
re-create a navigation structure based on that query.

There were also previous and next links that were doing more juggling
of that same data structure, not always with the greatest result.

All of this has been moved to the build script (gatsby-node.js) and
the required data is now injected into the page context. Specifically
for showcase and blog posts:

 - slug
 - title
 - next
 - previous

In addition, for docs:

 - offspring (children is a reserved prop name in react)
 - siblings
 - crumbs

Now, navigation is built using the page context, so there's no more
navigation menu that does all the processing at run time.
Same thing for crumbs.

We also have a new next/previous component that uses the same data.

The side effect of this is that we're funneling less data around.
The only downside is that we no longer have a 'full' overview of the
entire documentation structure anywhere. But we don't need it since
the navigation works different now.

One exception is the `sitemap` page which is now also created
by the build scripts script and receives the entire docs tree in
the page context.

**Context**

The new side navigation and the new `default` layout show a *context*
that is relevant to the page.
Rather than funneling this object around, we set it in the app object
in a `useEffect` call since we already use that to set title/description
which triggers a re-render anyway.

**The `useApp` hook**

The object returned by the `useApp` hook is a bit like a god object
that holds everything required to make the site do its thing.

A lot of logic in the app hook is related to account management,
drafting patterns and so on. That kind of stuff is not required
on static pages like blog posts, showcase posts, documentation and
so on.

So now the `useApp` hook takes an argument that defaults to `true`.
If you set it to `false` you will get the strict minimum.
By default, you get the full app object. So:

 - `const app = useApp(false)` -> for documentation pages and so on
 - `const app = useApp()` -> for the web app

I still have to update the mobile menu, and there will be things to do
left and right, but I'm wrapping this up in a commit now because it's
already rather large.