chore: Add @sewanista as a contributor
This commit will add you to the contributors file.
The result is that you will be listed on the website as a contributor.
For this reason, you should approve this before this gets merged.

In other words, this commit is a paper trail for adding this contributor
to the site.